backend/pyproject.toml
[project]
name = "knowledge-assistant-backend"
version = "0.1.0"
description = "Personal Knowledge Assistant Backend API"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "apscheduler>=3.11.1",
    "asyncpg>=0.31.0",
    "beautifulsoup4==4.12.3",
    "black>=25.12.0",
    "chromadb>=1.3.7",
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.3",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "langchain-text-splitters>=1.1.0",
    "markdown>=3.10",
    "mypy>=1.19.1",
    "ollama>=0.6.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.38.0",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.12.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.0",
]
